Rework RadioGroup for desktop
- Consolidate to a single file
- Use Button component directly, remove hidden input
- Use icons instead of CSS for radio circle graphic
- Update styles for desktop theme
- Ensure tabbing through options works
